slothjs: .interpret() now returns `Interpreted`, not `JSValue`.
The previous approach of having `Interpretable` and `IntepretableMut`
was not successful: it turned out that it does not make much sense to
separate expressions into returning JSValue and returning JSRef.

In order to support complex object assignment, `Interpreted` also needs
to support `Interpreted::Member{ of, name }`.

TODO: the JSValue::Object variant does not make sense too. A `JSValue`
must have a JSValue::Ref variant. The heap should be made of `JSObject`s,
not `JSValue`s.