slothjs: initial closures support
Make closures work in a brutal, but effective way: just capture
the whole scope into `captured_scope` of `Closure` when interpreting
a FunctionExpression.

When CallExpression is executed, `captured_scope` is put into the
new local scope as `[[captured_scope]]` pseudovariable. Variable lookups
now unwind the chain of captured scopes, if any.

In the future: try the approach of computing free variables upfront and
capture this set of references only. Compare this to the current savage
approach.