slothjs: variable hoisting
There is this interesting behavior:

    (function() {       // a is created here, with value `undefined`
        return a;       // return `undefined`
        var a = 'not gonna assign'; // but parser is aware of `var a`
    )()

The same behavior happens for function declarations.

So, a JavaScript parser cannot naively create variables when it executes
the `var` line: it has to look forward for all variable & function
declarations in a scope, then it has to stash them. When a closure is
executed, all local variables are created before executing anything,
they are just set to `undefined`, so runtime execution of `var <var> =
<val>` only assign the "initializing" value. This is called _variable
hoisting_ [0]

Also, now the parser knows the exact set of free variables for each
FunctionExpression, which will be handy in the future to optimize the
blunderbuss of `CAPTURED_SCOPE`.

[0] https://developer.mozilla.org/en-US/docs/Glossary/Hoisting