slothjs: optimization: don't copy Closure on every call
Previously, Closure was copied on every call (a dumb workaround for
borrow checker). After measuring things in callgrind, it became obvious
that copying the closure body on every call is a bad idea. Now a
JSObject::value contains not a `Box<Closure>`, but `Rc<Closure>`, which
means cloning only the Rc-pointer.

Program::interpret() execution time:
- `valgrind --tool=callgrind sljson.old < tmp/esprima.json`: before: 7.20%
- `valgrind --tool=callgrind sljson < tmp/esrpima.json`:     after: 3.65%

Program::parse_from() at 58% of the program execution time, which is not
surprising for the entire `esprima.json`, but this still leaves much room
for improvement.